requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version <= "3.11.2"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4.0"
cfgv==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
email-validator==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
fastapi[all]==0.95.1 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.12.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.24.0 ; python_version >= "3.10" and python_version < "4.0"
identify==2.5.24 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mongoengine==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
mongomock==4.1.2 ; python_version >= "3.10" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.8.10 ; python_version >= "3.10" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.10" and python_version < "4.0"
pinject==0.14.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==3.5.0 ; python_version >= "3.10" and python_version < "4.0"
pre-commit==3.3.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic-collections==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.6.0 ; python_version >= "3.10" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==4.5.4 ; python_version >= "3.10" and python_version < "4.0"
sentinels==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.7.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.26.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
ujson==5.7.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
virtualenv==20.23.0 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.1 ; python_version >= "3.10" and python_version < "4.0"
